Refactor: Use hooks on summert sykefravær
